chore: move visitors and implementation in the same source file
